Created some new components and renamed others to help with design
