+++ client/src/Components/EventHomeItem.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { ListItem, ListItemTest, Typography, Paper } from "@material-ui/core";
+import { makeStyles } from "@material-ui/core/styles"; 
+
+
+
+function EventHomeItem(props) {
+  return (
+    <ListItem>
+
+    </ListItem>
+  )
+}
+
+export default EventHomeItem;
