+++ client/src/Components/OrgList.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function OrgList(props) {
+  return(
+    <div></div>
+  );
+}
+
+export default OrgList;
